moderators done

+++ backend/src/services/moderator.service.js
@@ -1,5 +1,6 @@
 const TopicModerator = require('../models/TopicModerator.model');
 const TopicBlock = require('../models/TopicBlock.model');
+const Topic = require('../models/Topic.model');
 const User = require('../models/User.model');
 
 const {
@@ -11,9 +12,30 @@ const {
 
 async function addModerator(userId, topicId, targetUserId){
     const user = await User.findById(userId);
+    const topic = await Topic.findById(topicId);
 
-    if(!(await isModerator(userId, topicId)) || user.role === 'ADMIN'){
+
+    if(topic.isClosed){
+        throw new Error('Topic is closed');
+    }
+
+    if(topic.isHidden){
+        throw new Error('Topic is hidden');
+    }
+
+
+    const targetUser = await User.findById(targetUserId);
+    if (!targetUser){
+        throw new Error('User not found')
+
+    } 
+
+    if(!(await isModerator(userId, topicId)) && user.role !== 'ADMIN'){
         throw new Error('Not a moderator or admin');
+    }
+
+    if(targetUser.status !== 'ACTIVE'){
+        throw new Error('User is banned or pending')
     }
 
     if(await isModerator(targetUserId, topicId)){
@@ -30,8 +52,23 @@ async function addModerator(userId, topicId, targetUserId){
 async function removeModerator(userId, topicId, targetUserId){
     const user = await User.findById(userId);
 
-    if(!(await isModerator(userId, topicId)) || user.role === 'ADMIN'){
+    const topic = await Topic.findById(topicId);
+
+    if(topic.isClosed){
+        throw new Error('Topic is closed');
+    }
+
+    if(topic.isHidden){
+        throw new Error('Topic is hidden');
+    }
+
+
+    if(!(await isModerator(userId, topicId)) && user.role !== 'ADMIN'){
         throw new Error('Not a moderator or admin');
+    }
+
+    if(String(userId) === String(targetUserId)){
+        throw new Error("Cannot remove moderator from yourself");
     }
 
     if(!(await isModerator(targetUserId, topicId))){
@@ -41,10 +78,15 @@ async function removeModerator(userId, topicId, targetUserId){
     if(await isTopicFounderInBranch(targetUserId, topicId)){
         throw new Error('Cannot remove moderator privileges from topic founder');
     }
+    
 
     //higher decision is more important
     const localMod = await TopicModerator.findOne({ topicId, userId: targetUserId });
     if (!localMod) throw new Error('Cannot remove moderating. User was granted on a higher topic.');
+
+    if (user.role !== 'ADMIN' && String(localMod.promotedBy) !== String(userId)) {
+        throw new Error('You can only remove moderators you promoted');
+    }
 
 
     await TopicModerator.deleteOne({
@@ -55,9 +97,32 @@ async function removeModerator(userId, topicId, targetUserId){
 
 async function blockUser(userId, topicId, targetUserId, exceptions = [], reason){
     const user = await User.findById(userId);
+    const targetUser = await User.findById(targetUserId);
 
-    if(!(await isModerator(userId, topicId)) || user.role === 'ADMIN'){
+    const topic = await Topic.findById(topicId);
+
+    if(topic.isClosed){
+        throw new Error('Topic is closed');
+    }
+
+    if(topic.isHidden){
+        throw new Error('Topic is hidden');
+    }
+
+    if (!targetUser){
+        throw new Error('User not found');
+    }
+
+    if(targetUser.role === 'ADMIN'){
+        throw new Error('Cannot block admin');
+    }
+
+    if(!(await isModerator(userId, topicId)) && user.role !== 'ADMIN'){
         throw new Error('Not a moderator or admin');
+    }
+
+    if(String(userId) === String(targetUserId)){
+        throw new Error("Cannot block yourself");
     }
 
 
@@ -81,10 +146,23 @@ async function blockUser(userId, topicId, targetUserId, exceptions = [], reason)
 async function unblockUser(userId, topicId, targetUserId) {
     const user = await User.findById(userId);
 
-    if(!(await isModerator(userId, topicId)) || user.role === 'ADMIN'){
+    const topic = await Topic.findById(topicId);
+
+    if(topic.isClosed){
+        throw new Error('Topic is closed');
+    }
+
+    if(topic.isHidden){
+        throw new Error('Topic is hidden');
+    }
+
+    if(!(await isModerator(userId, topicId)) && user.role !== 'ADMIN'){
         throw new Error('Not a moderator or admin');
     }
 
+    if(String(userId) === String(targetUserId)){
+        throw new Error("Cannot unblock yourself");
+    }
 
     if (!(await isUserBlocked(targetUserId, topicId))) {
         throw new Error('User is not blocked');
